refactor(auth): hooks 폴더명을 model로 변경하고 useLoginStore를 shared에서 features/auth/model로 이동
