+++ src/app/(house)/layout.tsx
@@ -5,7 +5,7 @@ import { usePathname } from "next/navigation";
 import { NavigationBar } from "@/shared/components/NavigationBar";
 import { NAVIGATION_BAR_HEIGHT } from "@/shared/components/NavigationBar/style.css";
 import { getMe } from "@/shared/api/getMe";
-import { useLoginStore } from "@/shared/store/login";
+import { useLoginStore } from "@/features/auth/model/useLoginStore";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   const pathname = usePathname();
