+++ src/features/auth/components/Logout/index.tsx
@@ -1,5 +1,5 @@
 import { LogoutButton } from "@/features/auth/components/Logout/LogoutButton";
-import { usePostLogout } from "@/features/auth/hooks/usePostLogout";
+import { usePostLogout } from "@/features/auth/model/usePostLogout";
 
 export const Logout = () => {
   const { mutate: logout } = usePostLogout();
