+++ src/shared/components/NavigationBar/index.tsx
@@ -3,9 +3,9 @@
 import Link from "next/link";
 import Logo from "@/shared/components/NavigationBar/logo.svg";
 import { SearchForm } from "@/shared/components/NavigationBar/SearchForm";
-import { useLoginStore } from "@/shared/store/login";
-import * as s from "./style.css";
+import { useLoginStore } from "@/features/auth/model/useLoginStore";
 import { Logout } from "@/features/auth/components/Logout";
+import * as s from "./style.css";
 
 export const NavigationBar = () => {
   const { isLoggedIn } = useLoginStore();
